requirements.txt
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fonttools==4.60.1
-e git+ssh://git@github.com/Larraguibel/historical_figures_KG.git@42666ea700bb2201d4d039c053ef12e6231b920a#egg=historical_figures_kg
html2text==2025.4.15
html5rdf==1.2.1
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
owlrl==7.1.4
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prettytable==3.16.0
prompt_toolkit==3.0.52
psutil==7.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycurl==7.45.7
Pygments==2.19.2
pyparsing==3.2.5
pyshacl==0.30.1
python-dateutil==2.9.0.post0
pyvis==0.3.2
PyYAML==6.0.3
pyzmq==27.1.0
rdflib==7.2.1
requests==2.32.5
six==1.17.0
soupsieve==2.8
SPARQLWrapper==2.0.0
stack-data==0.6.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
wptools==0.4.17
